requirements.txt
mysql-connector-python
pandas
matplotlib
seaborn